services/ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
spacy==3.7.2
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.26.3
